requirements.txt
django==3.0.4
django-import-export==2.0.2
python-crontab==2.4.0